+++ src/api/client.ts
@@ -1,7 +1,13 @@
 import { decode, encode } from "@msgpack/msgpack";
-import Websocket from "isomorphic-ws";
 import Emittery from "emittery";
-import { AppSignal, AppSignalCb, SignalResponseGeneric } from "./app/types.js";
+import Websocket from "isomorphic-ws";
+import { AppSignal, Signal, SignalType } from "./app/types.js";
+
+interface HolochainMessage {
+  id: number;
+  type: "response" | "signal";
+  data: ArrayLike<number> | null;
+}
 
 /**
  * A Websocket client which can make requests and receive responses,
@@ -14,63 +20,89 @@ export class WsClient extends Emittery {
   pendingRequests: Record<
     number,
     {
-      fulfill: (msg: unknown) => ReturnType<typeof decode>;
+      resolve: (msg: unknown) => ReturnType<typeof decode>;
       reject: (error: Error) => void;
     }
   >;
   index: number;
 
-  constructor(socket: any, signalCb?: AppSignalCb) {
+  constructor(socket: Websocket) {
     super();
     this.socket = socket;
     this.pendingRequests = {};
     this.index = 0;
 
-    if (signalCb) {
-      console.log(
-        "Providing a signal callback on client initialization is deprecated. Instead add an event handler using `.on('signal', signalCb)`."
-      );
-      this.on("signal", signalCb);
-    }
-
-    socket.onmessage = async (encodedMsg: any) => {
-      let data = encodedMsg.data;
-
+    socket.onmessage = async (serializedMessage) => {
       // If data is not a buffer (nodejs), it will be a blob (browser)
-      if (typeof Buffer === "undefined" || !Buffer.isBuffer(data)) {
-        data = await data.arrayBuffer();
+      let deserializedData;
+      if (serializedMessage.data instanceof Blob) {
+        deserializedData = await serializedMessage.data.arrayBuffer();
+      } else {
+        if (
+          typeof Buffer !== "undefined" &&
+          Buffer.isBuffer(serializedMessage.data)
+        ) {
+          deserializedData = serializedMessage.data;
+        } else {
+          throw new Error("websocket client: unknown message format");
+        }
       }
 
-      const msg: any = decode(data);
-      if (msg.type === "signal") {
-        const decodedMessage: SignalResponseGeneric<any> = decode(msg.data);
+      const message = decode(deserializedData);
+      assertHolochainMessage(message);
 
-        if (!decodedMessage.App) {
+      if (message.type === "signal") {
+        if (message.data === null) {
+          throw new Error("received a signal without data");
+        }
+        const derializedSignal = decode(message.data);
+        assertHolochainSignal(derializedSignal);
+
+        if (!(SignalType.App in derializedSignal)) {
           // We have received a system signal, do nothing
           return;
         }
 
-        // Note: holochain currently returns signals as an array of two values: cellId and the serialized signal payload
+        // Holochain currently returns signals as an array of two values: cellId and the serialized signal payload
         // and this array is nested within the App key within the returned message.
-        const decodedCellId = decodedMessage.App[0];
-        // Note:In order to return readible content to the UI, the signal payload must also be decoded.
-        const decodedPayload = signalTransform(decodedMessage.App[1]);
+        const cell_id = derializedSignal.App[0];
+        // In order to return readible content to the UI, the signal payload must also be deserialized.
+        const payload = decode(derializedSignal.App[1]);
 
         // Return a uniform format to UI (ie: { type, data } - the same format as with callZome and appInfo...)
-        const signal: AppSignal = {
-          type: msg.type,
-          data: { cellId: decodedCellId, payload: decodedPayload },
-        };
+        const signal: AppSignal = { cell_id, payload };
         this.emit("signal", signal);
-      } else if (msg.type === "response") {
-        this.handleResponse(msg);
+      } else if (message.type === "response") {
+        this.handleResponse(message);
       } else {
-        console.error(`Got unrecognized Websocket message type: ${msg.type}`);
+        console.error(
+          `Got unrecognized Websocket message type: ${message.type}`
+        );
       }
     };
   }
 
-  emitSignal(data: any) {
+  static connect(url: string) {
+    return new Promise<WsClient>((resolve, reject) => {
+      const socket = new Websocket(url);
+      // make sure that there are no uncaught connection
+      // errors because that causes nodejs thread to crash
+      // with uncaught exception
+      socket.onerror = () => {
+        reject(
+          new Error(
+            `could not connect to holochain conductor, please check that a conductor service is running and available at ${url}`
+          )
+        );
+      };
+      socket.onopen = () => {
+        const client = new WsClient(socket);
+        resolve(client);
+      };
+    });
+  }
+
+  emitSignal(data: unknown) {
     const encodedMsg = encode({
       type: "signal",
       data: encode(data),
@@ -86,8 +118,8 @@ export class WsClient extends Emittery {
       type: "request",
       data: encode(data),
     });
-    const promise = new Promise((fulfill, reject) => {
-      this.pendingRequests[id] = { fulfill, reject };
+    const promise = new Promise((resolve, reject) => {
+      this.pendingRequests[id] = { resolve, reject };
     });
     if (this.socket.readyState === this.socket.OPEN) {
       this.socket.send(encodedMsg);
@@ -97,53 +129,52 @@ export class WsClient extends Emittery {
     return promise as Promise<Res>;
   }
 
-  handleResponse(msg: any) {
+  handleResponse(msg: HolochainMessage) {
     const id = msg.id;
     if (this.pendingRequests[id]) {
-      // resolve response
       if (msg.data === null || msg.data === undefined) {
         this.pendingRequests[id].reject(
           new Error("Response canceled by responder")
         );
       } else {
-        this.pendingRequests[id].fulfill(decode(msg.data));
+        this.pendingRequests[id].resolve(decode(msg.data));
       }
+      delete this.pendingRequests[id];
     } else {
       console.error(`Got response with no matching request. id=${id}`);
     }
   }
 
-  close(): Promise<void> {
+  close() {
+    const closedPromise = new Promise<void>((resolve) =>
+      this.socket.on("close", resolve)
+    );
     this.socket.close();
-    return this.awaitClose();
-  }
-
-  awaitClose(): Promise<void> {
-    return new Promise((resolve) => this.socket.on("close", resolve));
-  }
-
-  static connect(url: string, signalCb?: AppSignalCb): Promise<WsClient> {
-    return new Promise((resolve, reject) => {
-      const socket = new Websocket(url);
-      // make sure that there are no uncaught connection
-      // errors because that causes nodejs thread to crash
-      // with uncaught exception
-      socket.onerror = () => {
-        reject(
-          new Error(
-            `could not connect to holochain conductor, please check that a conductor service is running and available at ${url}`
-          )
-        );
-      };
-      socket.onopen = () => {
-        resolve(new WsClient(socket, signalCb));
-      };
-    });
+    return closedPromise;
   }
 }
 
-const signalTransform = (
-  res: SignalResponseGeneric<Buffer>
-): SignalResponseGeneric<any> => {
-  return decode(res);
-};
+function assertHolochainMessage(
+  message: unknown
+): asserts message is HolochainMessage {
+  if (
+    typeof message === "object" &&
+    message !== null &&
+    "type" in message &&
+    "data" in message
+  ) {
+    return;
+  }
+  throw new Error(`unknown message format ${JSON.stringify(message, null, 4)}`);
+}
+
+function assertHolochainSignal(signal: unknown): asserts signal is Signal {
+  if (
+    typeof signal === "object" &&
+    signal !== null &&
+    Object.values(SignalType).some((type) => type in signal)
+  ) {
+    return;
+  }
+  throw new Error(`unknown signal format ${JSON.stringify(signal, null, 4)}`);
+}